Add keybind remap UI, UI scale, cache guard, and infrastructure (#1)
- Keybindings: remappable playback controls via Settings with key-capture
  overlay; live re-resolved on save. New internal/keymap package with
  Name/Parse round-trip test.
- Display UI Scale: config field + Settings cycler (0.75–2.00) that
  shrinks the logical canvas so fonts/posters appear physically larger;
  zero-coerced and clamped.
- Image cache: buffer full response and decode in memory before writing
  to disk, so partial/undecodable downloads never persist. New test
  covers the unhappy path.
- Internal packages: httpx (retry with jitter + test), logging (slog
  setup), player/mpvcmd (IPC command builder + tests).
- Config: schema_version with migrate(), validate() with clamp/color
  warnings, atomic tmp+rename Save, UIScale coercion.
- UI: reusable Button helpers, clipboard paste button in text inputs,
  signout action, various screen cleanups.
- Tests: config round-trip/clamp/migrate, jellyfin client, image cache,
  http retry, keymap round-trip, mpvcmd.

+++ internal/ui/lang_editor.go
@@ -166,13 +166,13 @@ func (le *LangEditor) Update() {
 
 func (le *LangEditor) Draw(dst *ebiten.Image) {
 	// Semi-transparent overlay
-	vector.DrawFilledRect(dst, 0, 0, ScreenWidth, ScreenHeight, ColorOverlay, false)
+	vector.DrawFilledRect(dst, 0, 0, float32(ScreenWidth), float32(ScreenHeight), ColorOverlay, false)
 
 	// Centered panel
 	panelW := float32(1100)
 	panelH := float32(700)
-	panelX := float32(ScreenWidth-panelW) / 2
-	panelY := float32(ScreenHeight-panelH) / 2
+	panelX := (float32(ScreenWidth) - panelW) / 2
+	panelY := (float32(ScreenHeight) - panelH) / 2
 
 	vector.DrawFilledRect(dst, panelX, panelY, panelW, panelH, ColorBackground, false)
 	vector.StrokeRect(dst, panelX, panelY, panelW, panelH, 2, ColorPrimary, false)
